backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
Markdown==3.10.1
pydantic==2.12.5
pydantic-settings==2.12.0
python-multipart==0.0.22
kokoro==0.7.16
httpx==0.28.1
SQLAlchemy==2.0.46
aiosqlite==0.22.1
bcrypt==5.0.0
python-jose[cryptography]==3.5.0
